+++ x/btcstaking/types/mocked_keepers.go
@@ -180,17 +180,17 @@ func (m *MockBTCStkConsumerKeeper) EXPECT() *MockBTCStkConsumerKeeperMockRecorde
 }
 
 // ConsumerHasIBCChannelOpen mocks base method.
-func (m *MockBTCStkConsumerKeeper) ConsumerHasIBCChannelOpen(ctx context.Context, consumerID string) bool {
+func (m *MockBTCStkConsumerKeeper) ConsumerHasIBCChannelOpen(ctx context.Context, consumerID, channelID string) bool {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ConsumerHasIBCChannelOpen", ctx, consumerID)
+	ret := m.ctrl.Call(m, "ConsumerHasIBCChannelOpen", ctx, consumerID, channelID)
 	ret0, _ := ret[0].(bool)
 	return ret0
 }
 
 // ConsumerHasIBCChannelOpen indicates an expected call of ConsumerHasIBCChannelOpen.
-func (mr *MockBTCStkConsumerKeeperMockRecorder) ConsumerHasIBCChannelOpen(ctx, consumerID interface{}) *gomock.Call {
+func (mr *MockBTCStkConsumerKeeperMockRecorder) ConsumerHasIBCChannelOpen(ctx, consumerID, channelID interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ConsumerHasIBCChannelOpen", reflect.TypeOf((*MockBTCStkConsumerKeeper)(nil).ConsumerHasIBCChannelOpen), ctx, consumerID)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ConsumerHasIBCChannelOpen", reflect.TypeOf((*MockBTCStkConsumerKeeper)(nil).ConsumerHasIBCChannelOpen), ctx, consumerID, channelID)
 }
 
 // GetConsumerID mocks base method.